Exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -3,14 +3,6 @@ const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
-const getTokenFrom = request => {
-  const auth = request.get('authorization')
-  if (auth && auth.toLowerCase().startsWith('bearer ')) {
-    return auth.substring(7)
-  }
-  return null
-}
-
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({}).populate('user', { username: 1, name: 1 })
@@ -21,7 +13,7 @@ blogsRouter.post('/', async (request, response) => {
   const body = request.body
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
-    return response.status(401).json({ error: 'missing or invalid token'})
+    return response.status(401).json({ error: 'missing or invalid token' })
   }
 
   if (!body.title || !body.url) {
@@ -39,15 +31,28 @@ blogsRouter.post('/', async (request, response) => {
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
-    
+
     response.status(201).json(savedBlog)
   }
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const result = await Blog.findByIdAndRemove(request.params.id)
-  if (result) {
-    response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'missing or invalid token' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog) {
+    if (blog.user.toString() === decodedToken.id) {
+      const result = await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(401).json({
+        error: 'to delete a blog login with the creator\'s id'
+      })
+    }
   } else {
     response.status(404).end()
   }
